Card component viewport size improvements (#313)
* Added example bottom sheet containing the card component

* Changed viewport size update type to int and added bottom spacing to the viewport

* Replaced resize stream with state update

* Improved bottom sheet closing behavior

* extracted resize viewport into own method

* Added changelog entry

* Code formatting

* Moved default viewport spacing to Flutter layer in order to align with Android

+++ android/src/main/kotlin/com/adyen/checkout/flutter/components/view/DynamicComponentView.kt
@@ -122,17 +122,17 @@ class DynamicComponentView
             }
         }
 
-        private fun calculateFlutterViewportHeight(): Float {
+        private fun calculateFlutterViewportHeight(): Int {
             val componentViewHeightScreenDensity = measuredHeight / screenDensity
-            return componentViewHeightScreenDensity
+            return componentViewHeightScreenDensity.toInt()
         }
 
-        private fun resizeFlutterViewport(viewportHeight: Float) {
+        private fun resizeFlutterViewport(viewportHeight: Int) {
             componentFlutterApi?.onComponentCommunication(
                 ComponentCommunicationModel(
                     type = ComponentCommunicationType.RESIZE,
                     componentId = componentId,
-                    data = viewportHeight.toDouble()
+                    data = viewportHeight
                 )
             ) {}
         }
